requirements.txt
numpy==1.19.5
Pillow==8.2.0
matplotlib==3.4.1
tensorflow==2.5.0rc3
pandas==1.2.3
seaborn==0.11.1
sklearn==0.0
scikit-learn==0.24.1
imblearn==0.0
xgboost==1.4.2
catboost==0.26
ipython==7.23.0
opencv-contrib-python==4.5.2.52